performance issues
and redunant type casts

+++ QKSMS/src/main/java/com/moez/QKSMS/model/SlideshowModel.java
@@ -178,7 +178,7 @@ public class SlideshowModel extends Model
                             media.setDuration(mediadur);
                         }
 
-                        if ((int) mediadur / 1000 != dur) {
+                        if (mediadur / 1000 != dur) {
                             String tag = sme.getTagName();
 
                             if (ContentType.isVideoType(media.mContentType)
@@ -198,7 +198,7 @@ public class SlideshowModel extends Model
                                 * The Image disappear before the slide play done. so have to match
                                 * an image duration to the slide duration.
                                 */
-                                if ((int) mediadur / 1000 < dur) {
+                                if (mediadur / 1000 < dur) {
                                     media.setDuration((int) dur * 1000);
                                 } else {
                                     if ((int) dur != 0) {
