+++ QKSMS/src/main/java/com/moez/QKSMS/ui/view/slidingmenu/CustomViewAbove.java
@@ -275,7 +275,7 @@ public class CustomViewAbove extends ViewGroup {
     float distanceInfluenceForSnapDuration(float f) {
         f -= 0.5f; // center the values about 0.
         f *= 0.3f * Math.PI / 2.0f;
-        return (float) FloatMath.sin(f);
+        return FloatMath.sin(f);
     }
 
     public int getDestScrollX(int page) {
@@ -782,7 +782,7 @@ public class CustomViewAbove extends ViewGroup {
                 targetPage += 1;
             }
         } else {
-            targetPage = (int) Math.round(mCurItem + pageOffset);
+            targetPage = Math.round(mCurItem + pageOffset);
         }
         return targetPage;
     }
